Resolvers as static classes
