+++ Typin/Schemas/CommandParameterSchema.cs
@@ -3,7 +3,6 @@
     using System.Diagnostics.CodeAnalysis;
     using System.Reflection;
     using System.Text;
-    using Typin.Attributes;
 
     /// <summary>
     /// Stores command parameter schema.
@@ -20,36 +19,15 @@
         /// </summary>
         public string Name { get; }
 
-        #region ctor
         /// <summary>
         /// Initializes an instance of <see cref="CommandParameterSchema"/>.
         /// </summary>
-        private CommandParameterSchema(PropertyInfo? property, int order, string name, string? description)
+        internal CommandParameterSchema(PropertyInfo? property, int order, string name, string? description)
             : base(property, description)
         {
             Order = order;
             Name = name;
         }
-
-        /// <summary>
-        /// Resolves <see cref="CommandParameterSchema"/>.
-        /// </summary>
-        internal static CommandParameterSchema? TryResolve(PropertyInfo property)
-        {
-            CommandParameterAttribute? attribute = property.GetCustomAttribute<CommandParameterAttribute>();
-            if (attribute is null)
-                return null;
-
-            string name = attribute.Name ?? property.Name.ToLowerInvariant();
-
-            return new CommandParameterSchema(
-                property,
-                attribute.Order,
-                name,
-                attribute.Description
-            );
-        }
-        #endregion
 
         internal string GetUserFacingDisplayString()
         {
