--- Typin/Schemas/Resolvers/IResolver.cs
@@ -1,9 +1,0 @@
-﻿namespace Typin.Schemas.Resolvers
-{
-    internal interface IResolver<out T>
-        where T : class
-    {
-        bool IsResolved { get; }
-        T Resolve();
-    }
-}
